# Fixed log performance