Fixes #123 
``` 
                       │   0.bench    │                1.bench                │
                       │    sec/op    │    sec/op     vs base                 │
LongCity-32                2.036 ± ∞ ¹    1.893 ± ∞ ¹        ~ (p=1.000 n=1) ²
ManySmallCities-32        80.57µ ± ∞ ¹   48.12µ ± ∞ ¹        ~ (p=1.000 n=1) ²
LongField-32               2.052 ± ∞ ¹    2.313 ± ∞ ¹        ~ (p=1.000 n=1) ²
ManySmallFields-32        68.22µ ± ∞ ¹   37.37µ ± ∞ ¹        ~ (p=1.000 n=1) ²
LongMonastery-32           2.199 ± ∞ ¹    2.002 ± ∞ ¹        ~ (p=1.000 n=1) ²
ManySmallMonasteries-32   56.18µ ± ∞ ¹   35.97µ ± ∞ ¹        ~ (p=1.000 n=1) ²
SingleExtraLongRoad-32     2.060 ± ∞ ¹    2.073 ± ∞ ¹        ~ (p=1.000 n=1) ²
ManyShortRoads-32         59.45µ ± ∞ ¹   36.32µ ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                   11.69m         8.991m        -23.06%
```